use triggers to auto remove orphaned transactions
